+++ doc/release-notes.md
@@ -4,4 +4,8 @@ Bitcoin ABC version 0.32.12 is now available from:
 
   <https://download.bitcoinabc.org/0.32.12/>
 
-This is a maintenance release with no user-visible change.
+- If `-proxy=` is given together with `-noonion` then the provided proxy will
+  not be set as a proxy for reaching the Tor network. So it will not be
+  possible to open manual connections to the Tor network for example with the
+  `addnode` RPC. To mimic the old behavior use `-proxy=` together with
+  `-onlynet=` listing all relevant networks except `onion`.